Adding types to stream attribs, moving entry+instrument into group.py
