--- nexus_constructor/model/entry.py
@@ -1,6 +1,0 @@
-from nexus_constructor.model.group import Group
-
-
-class Entry(Group):
-    def __init__(self):
-        self.nx_class = "NXentry"
